requirements.txt
requests
python-dotenv
streamlit
feedparser
newsapi-python
sqlalchemy
beautifulsoup4
transformers
torch
pandas
huggingface_hub
google-generativeai
tabulate
watchdog